+++ funcverbnet/__init__.py
@@ -2,4 +2,4 @@
 
 __author__ = """Software Engineering Laboratory of Fudan University"""
 __email__ = 'lmwtclmwtc@outlook.com'
-__version__ = '0.2.2'
+__version__ = '0.2.6'